feat: add complement_alg_sd_inductive algorithm implementation and include in tests
