+++ src/algorithms/CMakeLists.txt
@@ -8,6 +8,7 @@ set(ALGORITHMS_SOURCES
     # complement_alg_rank.cpp  # Temporarily excluded due to compilation issues
     complement_alg_rank2.cpp
     complement_alg_safra.cpp
+    complement_alg_sd_inductive.cpp
     complement_alg_sd_tela.cpp
     complement_alg_subs_tuple.cpp
 )
@@ -21,6 +22,7 @@ set(ALGORITHMS_HEADERS
     complement_alg_rank.hpp
     complement_alg_rank2.hpp
     complement_alg_safra.hpp
+    complement_alg_sd_inductive.hpp
     complement_alg_sd_tela.hpp
     complement_alg_subs_tuple.hpp
     complement_mh.hpp
